neatogen scan_graph_mode: remove unhelpful comment
This note about a reader’s confusion is not helpful in actually understanding
the reason behind this allocation. AIUI, the reasoning is that this collection
is used to store ≤ the total number of nodes and most iteration through it uses
a null pointer check to detect the end. I.e.

  GD_neato_nlist(g) = gv_calloc(agnnodes(g), sizeof(node_t *));
  … append up to `agnnodes(g)` nodes …
  for (node_t **n = GD_neato_nlist(g); *n; ++n)
    …

Hence you need one extra entry for this null terminator.